Snapshot.cs and RepairSnapshot.cs
Added comments to Snapshot.cs for 'MergePDF' workaround to remove security from generated PDF file.
Created 'RepairSnapshot.cs' and added to project.  This class has a public function, 'ProcessFile' which checks for a corrupted DEVMODE structure in the Uncompressed Snapshot file and attempts to fix it prior to PDF conversion.  This fixes the Landscape/Cut-off page problem on converted PDFs.